Minor refinement in homepage explore gateway layout
