+++ src/components/homepage/ExploreGateway.js
@@ -4,7 +4,7 @@
  *
  * Author   : Tomiko
  * Created  : Jul 02, 2020
- * Updated  : Jul 28, 2020
+ * Updated  : Aug 08, 2021
  */
 
 import React from 'react'
@@ -13,13 +13,14 @@ import "semantic-ui-css/semantic.min.css"
 
 import ExploreGatewayIntro from './ExploreGatewayIntro'
 import ExploreGatewayGrid from './ExploreGatewayGrid'
+import ExploreGatewayFooter from './ExploreGatewayFooter'
 
 export default function ExploreGateway() {
   return (
     <div id="Explore" data-testid="ExploreGatewayComponentTestId">
       <ExploreGatewayIntro/>
       <ExploreGatewayGrid/>
-      <br/>
+      <ExploreGatewayFooter/>
     </div>
   );
 }
