+++ src/components/homepage/ExploreGatewayGrid.js
@@ -4,7 +4,7 @@
  *
  * Author   : Tomiko
  * Created  : Jul 02, 2020
- * Updated  : Aug 18, 2020
+ * Updated  : Aug 08, 2021
  */
 
 import React from 'react'
@@ -12,6 +12,8 @@ import React from 'react'
 import { Grid } from 'semantic-ui-react'
 
 import 'semantic-ui-css/semantic.min.css'
+
+import './ExploreGatewayGrid.css'
 
 import ExploreGatewayGridItem from './ExploreGatewayGridItem'
 
@@ -22,7 +24,7 @@ import explore_gateway_grid_item_img_04 from './assets/explore_gateway_grid_item
 
 export default function ExploreGatewayGrid() {
   return (
-    <nav>
+    <nav id="ExploreGatewayMenu" className="ExploreGatewayGrid">
       <Grid container stackable textAlign="center" columns={4} data-testid="ExploreGatewayGridTestId">
         <ExploreGatewayGridItem image={explore_gateway_grid_item_img_01} title="History" dstUrl="history" />
         <ExploreGatewayGridItem image={explore_gateway_grid_item_img_02} title="Biology" dstUrl="biology" />
