+++ src/components/homepage/ExploreGatewayIntro.css
@@ -7,9 +7,8 @@
  * Updated  : Aug 08, 2021
  */
 
-div.ExploreGatewayIntro {
-  background-color: rgba(239, 239, 239, 1.0);
-}
+/* div.ExploreGatewayIntro {
+} */
 
 p.JoinTheRace {
   font-size: 24px;
