+++ src/components/homepage/ExploreGatewayIntro.js
@@ -10,14 +10,10 @@
 import React from 'react'
 
 import {
-  Button,
   Container,
   Header,
   Icon,
-  Segment
 } from 'semantic-ui-react'
-
-import FluidSingleRowGrid from '../shared/FluidSingleRowGrid'
 
 import 'semantic-ui-css/semantic.min.css'
 
@@ -25,13 +21,15 @@ import '../shared/GlobalStyles.css'
 
 import '../shared/ContentPageSharedStyles.css'
 
+import './ExploreGatewaySharedStyles.css'
+
 import './ExploreGatewayIntro.css'
 
 import image_CCF_Cheetah_Museum from './assets/CCF_Cheetah_Museum-min.jpg'
 
 export default function ExploreGatewayIntro() {
   return (
-    <div className="ExploreGatewayIntro" data-testid="ExploreGatewayIntroTestId">
+    <div className="ExploreGatewayIntro ExploreGatewayBackground" data-testid="ExploreGatewayIntroTestId">
       <div className="VerticalCushionPadding">
         <img
           className="ExploreGatewayIntroCCFCheetahMuseumImage"
@@ -53,31 +51,6 @@ export default function ExploreGatewayIntro() {
             </p>
           </div>
         </Container>
-      </div>
-
-      <div id="ExploreGatewayMenu" className="ExploreGatewayIntroCenterGrid VerticalCushionPadding">
-        <FluidSingleRowGrid justifyContent="center">
-          <div className="ExploreGatewayIntroCenterGridLhsColumnStyle">
-            <p className="CCFMissionStatementQuote">
-              “CCF is Changing the World to Save the Cheetah” - Dr. Laurie Marker
-            </p>
-          </div>
-          <div className="ExploreGatewayIntroCenterGridRhsColumnStyle">
-            <Segment inverted>
-              <p className="JoinTheRace">
-                Join the race to <a href="https://twitter.com/search?q=%23SaveTheCheetah" target="_blank" rel="noopener noreferrer">
-                  <span className="SafeTheCheetah">
-                    #SafeTheCheetah
-                  </span>
-                </a>
-              </p>
-              <p>Cheetahs can't win without us.</p>
-              <Button basic inverted href="https://cheetah.org/get-involved/ways-to-give/">
-                Get Involved
-              </Button>
-            </Segment>
-          </div>
-        </FluidSingleRowGrid>
       </div>
     </div>
   );
